requirements.txt
Flask
mysql-connector-python
qrcode
reportlab
Flask-WTF
python-dotenv